% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaNB-package.R
\docType{package}
\name{betaNB-package}
\alias{betaNB}
\alias{betaNB-package}
\title{betaNB: Bootstrap for Regression Effect Sizes}
\description{
Generates nonparametric bootstrap confidence intervals (Efron and Tibshirani, 1993: \doi{10.1201/9780429246593}) for standardized regression coefficients (beta) and other effect sizes, including multiple correlation, semipartial correlations, improvement in R-squared, squared partial correlations, and differences in standardized regression coefficients, for models fitted by lm().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/betaNB}
  \item \url{https://jeksterslab.github.io/betaNB/}
  \item Report bugs at \url{https://github.com/jeksterslab/betaNB/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
