% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaNB-beta-nb.R
\name{BetaNB}
\alias{BetaNB}
\title{Estimate Standardized Regression Coefficients
and Generate the Corresponding Sampling Distribution
Using Nonparametric Bootstrapping}
\usage{
BetaNB(object)
}
\arguments{
\item{object}{Object of class \code{mc}, that is,
the output of the \code{MC()} function.}
}
\value{
Returns an object
of class \code{betanb} which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{object}{The function argument \code{object}.}
\item{thetahatstar}{Sampling distribution of
\eqn{\boldsymbol{\hat{\beta}}}.}
\item{jackknife}{Jackknife estimates.}
\item{est}{Vector of estimated
\eqn{\boldsymbol{\hat{\beta}}}.}
\item{fun}{Function used ("BetaNB").}
}
}
\description{
Estimate Standardized Regression Coefficients
and Generate the Corresponding Sampling Distribution
Using Nonparametric Bootstrapping
}
\details{
The vector of standardized regression coefficients
(\eqn{\boldsymbol{\hat{\beta}}})
is estimated from bootstrap samples.
Confidence intervals are generated by obtaining
percentiles corresponding to \eqn{100(1 - \alpha)\%}
from the generated sampling
distribution of \eqn{\boldsymbol{\hat{\beta}}},
where \eqn{\alpha} is the significance level.
}
\examples{
# Fit the regression model
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
# Generate the sampling distribution of sample covariances
# (use a large R, for example, R = 5000 for actual research)
nb <- NB(object, R = 50)
# Generate confidence intervals for standardized regression slopes
std <- BetaNB(nb)
# Methods --------------------------------------------------------
print(std)
summary(std)
coef(std)
vcov(std)
confint(std, level = 0.95)
}
\seealso{
Other Beta Nonparametric Bootstrap Functions: 
\code{\link{.CI}()},
\code{\link{DeltaRSqNB}()},
\code{\link{DiffBetaNB}()},
\code{\link{NB}()},
\code{\link{PCorNB}()},
\code{\link{RSqNB}()},
\code{\link{SCorNB}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Nonparametric Bootstrap Functions}
\keyword{betaNB}
\keyword{std}
