% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaNB-ci.R
\name{.CI}
\alias{.CI}
\title{Confidence Intervals}
\usage{
.CI(object, alpha = c(0.05, 0.01, 0.001), type = "pc")
}
\arguments{
\item{object}{Object of class \code{betanb}.}

\item{alpha}{Numeric vector.
Significance level.}

\item{type}{Charater string.
Confidence interval type, that is,
\code{type = "pc"} for percentile;
\code{type = "bc"} for bias corrected;
\code{type = "bca"} for bias corrected and accelerated.}
}
\value{
Returns a matrix of
estimates,
standard errors,
and
confidence intervals.
}
\description{
Confidence Intervals
}
\seealso{
Other Beta Nonparametric Bootstrap Functions: 
\code{\link{BetaNB}()},
\code{\link{DeltaRSqNB}()},
\code{\link{DiffBetaNB}()},
\code{\link{NB}()},
\code{\link{PCorNB}()},
\code{\link{RSqNB}()},
\code{\link{SCorNB}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Nonparametric Bootstrap Functions}
\keyword{betaNB}
\keyword{ci}
\keyword{internal}
