% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaNB-methods-beta-nb.R
\name{vcov.betanb}
\alias{vcov.betanb}
\title{Sampling Variance-Covariance Matrix Method for an Object of Class
\code{betanb}}
\usage{
\method{vcov}{betanb}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{betanb},
that is,
the output of the
\code{BetaNB()},
\code{RSqNB()},
\code{SCorNB()},
\code{DeltaRSqNB()},
\code{PCorNB()}, or
\code{DiffBetaNB()}
functions.}

\item{...}{additional arguments.}
}
\value{
Returns the variance-covariance matrix of estimates.
}
\description{
Sampling Variance-Covariance Matrix Method for an Object of Class
\code{betanb}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
