% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaNB-methods-nb.R
\name{print.nb}
\alias{print.nb}
\title{Print Method for an Object of Class \code{nb}}
\usage{
\method{print}{nb}(x, ...)
}
\arguments{
\item{x}{Object of Class \code{nb}.}

\item{...}{additional arguments.}
}
\value{
Prints the first six bootstrap covariance matrices.
}
\description{
Print Method for an Object of Class \code{nb}
}
\examples{
# Fit the regression model
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
# Generate bootstrap covariance matrices
# (use a large R, for example, R = 5000 for actual research)
nb <- NB(object, R = 50)
print(nb)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
