% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaNB-methods-beta-nb.R
\name{summary.betanb}
\alias{summary.betanb}
\title{Summary Method for an Object of Class
\code{betanb}}
\usage{
\method{summary}{betanb}(object, alpha = NULL, type = "pc", digits = 4, ...)
}
\arguments{
\item{object}{Object of Class \code{betanb},
that is,
the output of the
\code{BetaNB()},
\code{RSqNB()},
\code{SCorNB()},
\code{DeltaRSqNB()},
\code{PCorNB()}, or
\code{DiffBetaNB()}
functions.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.
If \code{alpha = NULL},
use the argument \code{alpha} used in \code{object}.}

\item{type}{Charater string.
Confidence interval type, that is,
\code{type = "pc"} for percentile;
\code{type = "bc"} for bias corrected;
\code{type = "bca"} for bias corrected and accelerated.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
estimates,
standard errors,
number of bootstrap replications,
and
confidence intervals.
}
\description{
Summary Method for an Object of Class
\code{betanb}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
